roles para mostrar opciones administrador y opciones de modificar

+++ public/js/incidence/incidences.js
@@ -228,9 +228,11 @@ $(function(){
 
             this.btn_delete = document.getElementById('btn_delete');
             this.btn_edit = document.getElementById('btn_edit');
+            this.btn_guardar = document.getElementById('btn_guardar');
             this.select_area = document.getElementById('area');
             this.select_encuesta = document.getElementById('encuesta');
             this.$select_categorias = $('#categories');
+            this.descripcion = document.getElementById('solicitude_description')
 
             this.render();
 
@@ -241,6 +243,44 @@ $(function(){
             console.log(this.select_area, this.btn_delete);
 
             this.$select_categorias.select2();
+
+            if(this.btn_guardar !== null){
+
+                _this = this;
+
+                this.btn_guardar.onclick = function(){
+
+                    $.post({
+                        url: document.getElementById('update_solicitude_route').value,
+                        data: {
+                            area: _this.select_area.value,
+                            encuesta: _this.select_encuesta.value,
+                            categorias: _this.$select_categorias.val(),
+                            descripcion: _this.descripcion.value
+                        },
+                        headers: {
+                            'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
+                        },
+                        success: function(response) {
+                            console.log(response);
+
+                            swal(
+                                'Exito',
+                                "Los cambios se han aplicado con exito",
+                                'success'
+                            );
+
+                        },
+                        error: function(response) {
+                            console.log('error', response);
+                        }
+                    });
+
+                }
+
+            }
+
+            /*
 
             this.select_area.onchange = function(){
 
@@ -310,48 +350,54 @@ $(function(){
                 });
 
             });
+
+            */
             
-            this.btn_delete.onclick = function(e){
+            if(this.btn_delete !== null){
 
-                e.preventDefault();
+                this.btn_delete.onclick = function(e){
 
-                _this = this;
-                
-                swal({
-                    title: '¿Esta seguro de eliminar esta solicitud?',
-                    type: 'warning',
-                    showCancelButton: true,
-                    confirmButtonText: 'SI',
-                    cancelButtonText: 'NO',
+                    e.preventDefault();
+    
+                    _this = this;
                     
-                }).then(function(isConfirm){
-                    
-                    if(isConfirm.value){
+                    swal({
+                        title: '¿Esta seguro de eliminar esta solicitud?',
+                        type: 'warning',
+                        showCancelButton: true,
+                        confirmButtonText: 'SI',
+                        cancelButtonText: 'NO',
+                        
+                    }).then(function(isConfirm){
+                        
+                        if(isConfirm.value){
+    
+                            //return;
+    
+                            $.ajax({
+    
+                                url: _this.href,
+                                type: 'DELETE',
+                                data: {
+                                    _method: 'delete',
+                                    _token: $('meta[name="csrf-token"]').attr('content')
+                                },
+                                success: function(response){
+                                    window.location = 'http://www.helpdesk.local/incidence/solicitudes';
+                                },
+                                error: function(response){
+                                    console.log(response);
+                                }
+            
+                            });
+    
+                        }
+    
+                    });
+    
+                };
 
-                        //return;
-
-                        $.ajax({
-
-                            url: _this.href,
-                            type: 'DELETE',
-                            data: {
-                                _method: 'delete',
-                                _token: $('meta[name="csrf-token"]').attr('content')
-                            },
-                            success: function(response){
-                                window.location = 'http://www.helpdesk.local/incidence/solicitudes';
-                            },
-                            error: function(response){
-                                console.log(response);
-                            }
-        
-                        });
-
-                    }
-
-                });
-
-            };
+            }
 
         }
